requirements.txt
fastapi
uvicorn
python-dotenv
pymysql
sqlalchemy
pandas 